chore(web): remove dead routes/shims and relocate cancel-card
