--- apps/web/src/app/tasks/actions.ts
@@ -1,5 +1,0 @@
-export {
-  loadMoreJobs,
-  loadMoreTasksColumn,
-  loadMoreTasksList,
-} from "../(app)/tasks/actions";
